feat(aws): update credentials logic to hook into portainer helper
