+++ edge/edge.go
@@ -38,7 +38,7 @@ type (
 		ClusterService    agent.ClusterService
 		DockerInfoService agent.DockerInfoService
 		ContainerPlatform agent.ContainerPlatform
-		AWSConfig         *agent.AWSConfig
+		// AWSConfig         *agent.AWSConfig
 	}
 )
 
@@ -46,9 +46,9 @@ func (manager *Manager) GetStackManager() *stack.StackManager {
 	return manager.stackManager
 }
 
-func (manager *Manager) GetAWSConfig() *agent.AWSConfig {
-	return manager.awsConfig
-}
+// func (manager *Manager) GetAWSConfig() *agent.AWSConfig {
+// 	return manager.awsConfig
+// }
 
 // NewManager returns a pointer to a new instance of Manager
 func NewManager(parameters *ManagerParameters) *Manager {
@@ -58,7 +58,7 @@ func NewManager(parameters *ManagerParameters) *Manager {
 		agentOptions:      parameters.Options,
 		advertiseAddr:     parameters.AdvertiseAddr,
 		containerPlatform: parameters.ContainerPlatform,
-		awsConfig:         parameters.AWSConfig,
+		// awsConfig:         parameters.AWSConfig,
 	}
 }
